Add initial stage for test creation and ensure correct navigation back to the personal account

+++ app/users/user/userHandlers.py
@@ -1,3 +1,4 @@
+from aiogram.enums import ParseMode
 from aiogram.types import Message, CallbackQuery
 from aiogram.filters import CommandStart, Command
 from aiogram import F, Router
@@ -22,7 +23,8 @@ async def delete_previous_messages(message: Message):
     # Delete all bot messages
     for msg_id in sent_message_add_screen_ids['bot_messages']:
         try:
-            await message.bot.delete_message(chat_id=message.chat.id, message_id=msg_id)
+            if msg_id != message.message_id:
+                await message.bot.delete_message(chat_id=message.chat.id, message_id=msg_id)
         except Exception as e:
             print(f"Не удалось удалить сообщение {msg_id}: {e}")
     sent_message_add_screen_ids['bot_messages'].clear()
@@ -30,22 +32,29 @@ async def delete_previous_messages(message: Message):
 # User's personal account
 async def user_account(message: Message, state: FSMContext):
     sent_message_add_screen_ids['user_messages'].append(message.message_id)
-    await delete_previous_messages(message)
 
-    user_tg_id = str(message.from_user.id)
+    user_tg_id = str(message.chat.id)
     language = await rq.get_user_language(user_tg_id)
     name = await rq.get_user_name(user_tg_id)
+
+    await delete_previous_messages(message)
+
+
 
     if language == 'ru':
         sent_message = await message.answer_photo(
             photo=utils.pictureOfUsersPersonalAccountRU,
-            caption=f'Привет, {name}',
-        reply_markup=kb.profile_button_ru)
+            caption=f'Привет, {name}'
+                    f'\n<a href="https://telegra.ph/lpshchzk-10-30">Как бот работает?</a> 👈',
+        reply_markup=kb.profile_button_ru,
+        parse_mode=ParseMode.HTML)
     else:
         sent_message = await message.answer_photo(
-            photo=utils.pictureOfUsersPersonalAccountKG,
-            caption=f'Салам, {name}',
-        reply_markup=kb.profile_button_kg)
+            photo=utils.pictureOfUsersPersonalAccountRU,
+            caption=f'Салам, {name}'
+                    f'\n<a href="https://telegra.ph/Bizdin-ORTga-dayardanuu-%D2%AFch%D2%AFn-Telegram-bot-kandaj-ishtejt-10-30">Бот кандай иштейт?</a> 👈',
+            reply_markup=kb.profile_button_kg,
+            parse_mode=ParseMode.HTML)
 
     sent_message_add_screen_ids['bot_messages'].append(sent_message.message_id)
 
@@ -63,3 +72,22 @@ async def photo_handler(message: Message):
     # Берем фотографию в самом большом разрешении и получаем ее ID
     photo_id = message.photo[-1].file_id
     await message.answer(f"ID вашей картинки: {photo_id}")
+
+# Handler for creating a question
+@router.callback_query(F.data == 'creat_test_ru')
+async def create_question(callback_query: CallbackQuery, state: FSMContext):
+    sent_message_add_screen_ids['user_messages'].append(callback_query.message.message_id)
+    await delete_previous_messages(callback_query.message)
+    sent_message = await callback_query.message.answer(text="Выберите предмет, по которому вы хотели бы создать вопрос.",
+                                                       reply_markup=kb.subjects_ru)
+    sent_message_add_screen_ids['bot_messages'].append(sent_message.message_id)
+
+
+@router.callback_query(F.data.in_(['to_home_ru', 'to_home_kg']))
+async def go_home_handler(callback_query: CallbackQuery, state: FSMContext):
+    # Добавление ID сообщения в список
+    sent_message_add_screen_ids['bot_messages'].append(callback_query.message.message_id)
+
+    # Вызов функции для отображения личного кабинета
+    await user_account(callback_query.message, state)
+
